refactor: move dp accounting to seperate file
